Remove unneeded static cast
Fix #2030.

+++ os_stub/debuglib/debuglib.c
@@ -26,8 +26,7 @@
 
 void libspdm_debug_assert(const char *file_name, size_t line_number, const char *description)
 {
-    printf("LIBSPDM_ASSERT: %s(%d): %s\n", file_name, (int32_t)(uint32_t)line_number,
-           description);
+    printf("LIBSPDM_ASSERT: %s(%zu): %s\n", file_name, line_number, description);
 
 #if (LIBSPDM_DEBUG_LIBSPDM_ASSERT_CONFIG == LIBSPDM_DEBUG_LIBSPDM_ASSERT_DEADLOOP)
     {
